Checking that we have credentials for amazon before attempting to interact with the cloud -- back to simulation mode.

+++ src/main/java/com/zuehlke/pgadmissions/services/lifecycle/helpers/DocumentServiceHelperDelete.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Component;
 
 import com.zuehlke.pgadmissions.exceptions.DeduplicationException;
 import com.zuehlke.pgadmissions.services.DocumentService;
+import com.zuehlke.pgadmissions.services.SystemService;
 
 @Component
 public class DocumentServiceHelperDelete extends AbstractServiceHelper {
@@ -19,11 +20,14 @@ public class DocumentServiceHelperDelete extends AbstractServiceHelper {
     @Autowired
     private DocumentService documentService;
     
+    @Autowired
+    private SystemService systemService;
+    
     @Override
     public void execute() throws DeduplicationException, IOException {
         DateTime baselineTime = new DateTime();
         documentService.deleteOrphanDocuments(baselineTime);
-        if (amazonOn) {
+        if (amazonOn && systemService.getSystem().isDocumentExportEnabled()) {
             documentService.deleteAmazonDocuments(baselineTime);
         }
     }
